Update Compute Library to 21.11 release
Updates the version of Compute Library to the 21.11
This release adds support for bf16 fast-maths-mode to additional
oneDNN primitives - inner product and matmul.

oneDNN 2.4 build is patched to expose this functionality to TensorFlow

TF_ENABLE_ONEDNN_OPTS flag (runtime selection of oneDNN backend) is
enabled on AArch64 for builds with `--config=mkl_aarch64` set.

Co-authored-by: Diana Bite <diana.bite@arm.com>