[XLA:GPU] Remove fast minmax from default on GPU
Fast minmax has confusing semantics inconsistent with TF, and is not actually
fast.

For:

```
HloModule test

max {
  p0 = f16[] parameter(0)
  p1 = f16[] parameter(1)
  ROOT add = f16[] maximum(p0, p1)
}

ENTRY entry {
  p = f16[10000000] parameter(0)
  constant = f16[] constant(0)
  ROOT out = f16[] reduce(p, constant), to_apply=max, dimensions={0}
}
```

with --xla_gpu_enable_fast_min_max (default) we get:

```
  long name: %out = f16[] reduce(f16[10000000]{0} %p, f16[] %constant), dimensions={0}, to_apply=%max
  total time: 49us
  bytes accessed: 20000004
  bandwidth: 403 Gb/s
  bandwidth roofline: 52%
  time wasted: 23us
```

and then for --xla_gpu_enable_fast_min_max=false we get:

```
  long name: %out = f16[] reduce(f16[10000000]{0} %p, f16[] %constant), dimensions={0}, to_apply=%max
  total time: 49us
  bytes accessed: 20000004
  bandwidth: 406 Gb/s
  bandwidth roofline: 52%
  time wasted: 23us
```

Moreover, this is for the microbenchmark which is as sensitive as it gets to
max implementation (vectorizable, does atomic max operation most of the time).

PiperOrigin-RevId: 416330132
Change-Id: Ieafec281cbfffa4f972a6ac52b77acbaeb48370f