[ROCm] Explicitly setting the `__launch_bounds__` attribute to 1024 for MLIR generated kernels (for ROCm only)
The following LLVM commit changes the default `__launch_bounds__` attribute from 1024 to 256 (for the ROCm path).

That results in quite a few TF unit-test failures, because TF side assumes/expects the default to be 1024, and launches GPU kernels at runtime with block-size == 1024. As a consequence, all unit-tests that do this, run into an error similar to the following error

```
...
:3:hip_module.cpp           :363 : 20798447993994 us: 2453 : [7f51477fe700] ihipModuleLaunchKernel ( 0x0x7f5138273b40, 1024, 1, 1, 1024, 1, 1, 0, stream:0x26b7fb0, 0x7f51477fcb30, char array:<null>, event:0, event:0, 0, 0 )
:1:hip_module.cpp           :256 : 20798447993999 us: Launch params (1024, 1, 1) are larger than launch bounds (256) for kernel Mul_GPU_DT_FLOAT_DT_FLOAT_kernel
...
```

The fix for this breakage requires changes on the both the TF side and LLVM side

On the LLVM side, we need to add the ability to configure/overwrite the default value (256) of the amdgpu-flat-work-group-size attribute for MLIR generated GPU kernels. This is done via the addition of the rocdl.max_flat_work_group_size attribute, which when set on the mlir::GPUFuncOp will override the default value (256) when mlir::GPUFuncOp gets translated to llvm::FuncOp. See the following LLVM PR for details

https://reviews.llvm.org/D115741

The above PR has already been merged, and the TF LLVM ptr has been updated to pick up that change.

On the TF side, we need to explicitly specify the rocdl.max_flat_work_group_size attribute with value set to 1024 for MLIR generated kernels...which is what this commit does.