Fix MHLO->HLO prepare pass to correctly handle when a value is captured multiple times in a while
We need to use make_early_inc_range to ensure that we pre-increment the iterator
before modifying it in the loop body. Before we were only updating the first use of a
value, and after the update the iterator pointed to the updated operand use-list and
the process would stop here.

PiperOrigin-RevId: 417718744
Change-Id: I4a2df523cdd3111fcebfb3e2b32800444a94d45b