Update double-conversion to 3.1.7, and use vendor-provided Bazel BUILD file instead.
The double-conversion library used in tensorflow has not been updated
for quite some time (previous version used was in 2018) and the Bazel
BUILD file was maintained by tensorflow.

Recently double-conversion has provided Bazel BUILD file.

This PR updates double-conversion to the latest 3.1.7, and use the vendor-provided
BUILD file instead so that tensorflow does not need to maintain the Bazel BUILD file
separately.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>