Fix HLO parser bug wrt attributes and nested instructions.
Nested instructions are not allowed to have attributes, because it's ambiguous
whether the next element is another operand or an attribute:

  add(broadcast(x), y)  // is y an operand or an attribute of the broadcast?

Our enforcement of this rule had the following bug.  If the nested op
*requires* some attributes (e.g. broadcast requires a `dimensions` attribute),
we would not check whether the attributes were present.  So e.g. we'd allow

  sqrt(f32[10,10] broadcast(x))

without any explicit broadcast dims, which is illegal.  This would ultimately
lead us to getting the value from a nullopt absl::optional.

PiperOrigin-RevId: 420623411
Change-Id: I791b97d40617811bcac9aff97ed192c446769f81