Treat TF_MustExecute as read effect in side effect analysis
Previously, we completely ignored this special effect in side effect analysis
which meant that after MLIR-to-graph conversion we didn't always have a
connection from an op with `TF_MustExecute` to the control output of the
surrounding function, meaning that the op would potentially not be executed.
This is now resolved by treating `TF_MustExecute` as read effect in side effect
analysis which ensures that we always have a dependency to some terminator op
(like `tf_executor.yield`) which results in a (transitive) control dependency
to the control output of the surrounding function after MLIR-to-graph
conversion.

PiperOrigin-RevId: 421346187
Change-Id: Ib071647bb2f8517c142af19799bc325f32a00217