Add cancellation support for cross-host transfers
CopyToRemoteDevice/CopyToRemoteDeviceScattered now take callbacks as arguments, and the callbacks fire when:
1) the send has not occurred because of an error (e.g., invalid parameter, or the send buffer transitioned to error)
2) the send was enqueued to the device but failed (e.g., because of a hardware error)
3) the send was enqueued to the device and completed successfully

Cases 1+2 are disambiguated using a bool in the callback.

The PjRtCrossHostReceiveNotifier that is used to create receive buffers for cross-host transfers now supplies a cancellation function that can be used to indicate that one of the transfers should be cancelled, e.g., because the remote send failed or the remote buffer is not available.

PiperOrigin-RevId: 421423275
Change-Id: I0b323274790897a7b3c1c1498c5b8c6ff37373d8