[tfrt:jit] Add jitrt CallingConvention to get the compiled function ABI from the entrypoint signature
Instead of two compilation pipelines (to-jitrt, to-llvm) we want to have a single compilation pipeline, and because of that we can't introspect that entrypoint signature in the middle of the pipeline.

Calling convention allows jitrt users to explicitly defined the expected ABI of the function that will be produced by the compilation pipeline.

PiperOrigin-RevId: 422520942
Change-Id: I1625cbb623f179e2d41b003dbe636444e26b45e4