Fix warning due to misuse of @available attribute.
The code was using

    @available(*, deprecated: 2.4, renamed: "isPrecisionLossAllowed")

but (probably as a result of a newer version of the Apple tools), this
now results in a warning

    unexpected version number in 'available' attribute for non-specific platform '*'

This warning is because it doesn't make sense to say that a feature is
deprecated in version 2.4 of *every* platform.  This code was misusing
the version number field, which is supposed to refer to a version of an
Apple platform (one of a fixed list: iOS, iOSApplicationExtension, macOS,
macOSApplicationExtension, macCatalyst, macCatalystApplicationExtension,
watchOS, watchOSApplicationExtension, tvOS, tvOSApplicationExtension,
or swift), and was instead trying to use it to refer to the version of
TF Lite.  The fix was to use the 'message' field instead.

PiperOrigin-RevId: 426504081
Change-Id: I8bf792b1973a96c4bb1cc747b118b02a80eb2985