[XLA:Python] Consistently use NPY_... types
NPY_... types live in two different places: npy_common.h and ndarraytypes.h
The types in npy_common.h are derived from the ones in ndarraytypes.h depending on the C ABI.

We are using a mixture of types from both which makes it non-obvious to determine which actual types are being used. Switching to the underlying enumerators from ndarraytypes.h makes it easy to see which ones are in use.

PiperOrigin-RevId: 426949188
Change-Id: I260cbbc0a76f235b39e9ccb4ebcf8c1e8b0cabd9